+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.68] - 2025-05-05
+
+### Added
+- Changes in CONTRIBUTING.md
+- Changes in README.md
+- Changes in _config.yml
+- Changes in docs/contributing_guide.md
+- Changes in docs/user_guide.md
+- Changes in pyproject.toml
+- Changes in src/twinizer.egg-info/PKG-INFO
+- Changes in src/twinizer.egg-info/SOURCES.txt
+- Changes in src/twinizer/__init__.py
+
 ## [0.1.67] - 2025-05-05
 
 ### Added